Bug fixes
This commit brings fixes to the bugs of the previous commit. It also changes the background image from a corkboard to a chalkboard. The side bar is still undeveloped as it's not too critical at the moment. Still hardcoding test data, will try to implement db communication later in a few more commits.