Added Search functionality
The user is now able to search for notes that match any of title/body/category.
Also deprecated the search button so the user can dynamically search as they type in the search bar.
Adjusted the timer in the LoadingScreen to 30s from 22s. For some reason,  the looping execution of getAllNotes() wasn't working properly, so instead of having to wait a couple more seconds after the timer has finished, the new time should more accurately reflect the actual delay.